refactor(students): finish api.go split (#575 B6)
Pure file moves: remaining core CRUD from api.go (1499 L → 198 L
wiring-only) into student_handlers.go; access checks join
authorization.go, response builders join response_helpers.go, the
renderError delegate joins errors.go. Route table unchanged incl. the
device-authenticated RFID group (B0 golden). Complexity ratchet entries
re-pathed, scores unchanged.
