+++ backend/api/students/student_handlers.go
@@ -1,0 +1,1130 @@
+package students
+
+import (
+	"context"
+	"database/sql"
+	"errors"
+	"fmt"
+	"log/slog"
+	"net/http"
+	"strings"
+	"time"
+
+	"github.com/go-chi/render"
+	"github.com/moto-nrw/project-phoenix/api/common"
+	guardiansAPI "github.com/moto-nrw/project-phoenix/api/guardians"
+	"github.com/moto-nrw/project-phoenix/auth/authorize"
+	"github.com/moto-nrw/project-phoenix/auth/authorize/permissions"
+	"github.com/moto-nrw/project-phoenix/auth/jwt"
+	"github.com/moto-nrw/project-phoenix/internal/strutil"
+	"github.com/moto-nrw/project-phoenix/internal/timezone"
+	configModel "github.com/moto-nrw/project-phoenix/models/config"
+	"github.com/moto-nrw/project-phoenix/models/users"
+	"github.com/moto-nrw/project-phoenix/realtime"
+	activeService "github.com/moto-nrw/project-phoenix/services/active"
+	configService "github.com/moto-nrw/project-phoenix/services/config"
+	userService "github.com/moto-nrw/project-phoenix/services/users"
+	"github.com/moto-nrw/project-phoenix/tenant"
+	"github.com/uptrace/bun"
+)
+
+// parseAndGetStudent parses the student ID from the URL and fetches the student
+// Returns the student and true if successful, or renders an error and returns nil, false
+func (rs *Resource) parseAndGetStudent(w http.ResponseWriter, r *http.Request) (*users.Student, bool) {
+	id, err := common.ParseID(r)
+	if err != nil {
+		renderError(w, r, common.ErrorInvalidRequest(errors.New(common.MsgInvalidStudentID)))
+		return nil, false
+	}
+
+	student, err := rs.PersonService.GetStudentByID(r.Context(), id)
+	if err != nil {
+		renderError(w, r, common.ErrorNotFound(errors.New("student not found")))
+		return nil, false
+	}
+
+	return student, true
+}
+
+// listStudents handles listing all students with staff-based filtering
+func (rs *Resource) listStudents(w http.ResponseWriter, r *http.Request) {
+	// Parse query parameters and determine access
+	params := parseStudentListParams(r)
+	accessCtx := rs.determineStudentAccess(r)
+
+	// Fetch students based on parameters
+	students, totalCount, err := rs.fetchStudentsForList(r, params)
+	if err != nil {
+		if errors.Is(err, ErrInvalidRequest) {
+			renderError(w, r, common.ErrorInvalidRequest(err))
+			return
+		}
+		renderError(w, r, common.ErrorInternalServer(err))
+		return
+	}
+
+	// Bulk load all related data
+	studentIDs, personIDs, groupIDs := collectIDsFromStudents(students)
+	dataSnapshot, err := common.LoadStudentDataSnapshot(
+		r.Context(),
+		rs.PersonService,
+		rs.EducationService,
+		rs.ActiveService,
+		studentIDs,
+		personIDs,
+		groupIDs,
+	)
+	if err != nil {
+		slog.Default().Error("failed to load student data snapshot", slog.String("error", err.Error()))
+		renderError(w, r, common.ErrorInternalServer(err))
+		return
+	}
+
+	// Resolve once per request. populatePhotoFields runs per student.
+	photosEnabled := configService.ResolveBoolOrDefault(r.Context(), rs.SettingsService, configModel.KeyStudentPhotosEnabled, false, rs.Logger)
+
+	// Build and filter responses
+	responses := rs.buildStudentResponses(r.Context(), students, params, accessCtx, dataSnapshot, photosEnabled)
+
+	now := rs.Now()
+	rs.applyStatusDaysForDate(r.Context(), responses, now)
+	if err := rs.enrichWithDayPlanning(r.Context(), responses, now, attendanceMapFromSnapshot(dataSnapshot)); err != nil {
+		slog.Default().Error("failed to enrich student day planning", slog.String("error", err.Error()))
+		renderError(w, r, common.ErrorInternalServer(err))
+		return
+	}
+	responses = applyDayPlanningFilter(responses, params.dayStatus)
+	// Administrative filters (#1492): bus / photo consent / pickup rule.
+	// Applied here, before in-memory pagination, so server-side counts and
+	// page boundaries reflect the filtered set (no client-side full-page
+	// fetch needed).
+	responses = applyAdministrativeFilters(responses, params.bus, params.photoConsent, params.pickupStatus)
+
+	// Apply in-memory pagination if response-derived filters were used.
+	if params.hasInMemoryFilters() {
+		responses, totalCount = applyInMemoryPagination(responses, params.page, params.pageSize)
+	}
+
+	for i := range responses {
+		if !responses[i].HasFullAccess {
+			continue
+		}
+		applyActualTimesFromSnapshot(&responses[i], dataSnapshot)
+	}
+
+	// Optionally enrich the paginated slice with today's effective pickup times (single bulk query).
+	// Only query for students the caller has full access to. GDPR: skip redacted students.
+	if params.includePickupTimes || params.includeArrivalTimes {
+		fullAccessIDs := collectFullAccessStudentIDs(responses)
+		if params.includePickupTimes {
+			rs.enrichWithPickupTimes(r.Context(), responses, fullAccessIDs, now)
+		}
+		if params.includeArrivalTimes {
+			rs.enrichWithArrivalTimes(r.Context(), responses, fullAccessIDs, now)
+		}
+	}
+
+	common.RespondPaginated(w, r, http.StatusOK, responses, common.PaginationParams{Page: params.page, PageSize: params.pageSize, Total: totalCount}, "Students retrieved successfully")
+}
+
+// fetchStudentsForList fetches students based on the provided parameters
+func (rs *Resource) fetchStudentsForList(r *http.Request, params *studentListParams) ([]*users.Student, int, error) {
+	ctx := r.Context()
+
+	if params.locationState != "" {
+		if params.locationState != "transit" && params.locationState != "present" {
+			return nil, 0, ErrInvalidRequest
+		}
+		if params.roomID > 0 {
+			return nil, 0, ErrInvalidRequest
+		}
+		var ids []int64
+		var err error
+		if params.locationState == "present" {
+			ids, err = rs.ActiveService.ListStudentsPresentToday(ctx)
+		} else {
+			ids, err = rs.ActiveService.ListStudentsInTransit(ctx)
+		}
+		if err != nil {
+			return nil, 0, err
+		}
+		if len(ids) == 0 {
+			return []*users.Student{}, 0, nil
+		}
+		if params.groupID > 0 {
+			ids, err = rs.filterStudentIDsByGroup(ctx, ids, params.groupID)
+			if err != nil {
+				return nil, 0, err
+			}
+			if len(ids) == 0 {
+				return []*users.Student{}, 0, nil
+			}
+		}
+		params.studentIDs = ids
+	}
+
+	// room_id pre-filter (#1323): resolve students currently checked-in to
+	// any active group in the room, then push the IDs through the standard
+	// query path so school_class / guardian_name / pagination still apply.
+	// The visit join lives in the active service (rule 11: services own
+	// queries, not handlers).
+	if params.roomID > 0 {
+		ids, err := rs.ActiveService.ListStudentsPresentInRoom(ctx, params.roomID)
+		if err != nil {
+			return nil, 0, err
+		}
+		if len(ids) == 0 {
+			return []*users.Student{}, 0, nil
+		}
+		// When both room_id and group_id are supplied, intersect with the
+		// student's group_id so the response stays consistent with the
+		// active-group chip in the search UI. group_id lives on the Student
+		// row, so a single bulk lookup is enough.
+		if params.groupID > 0 {
+			filtered, err := rs.filterStudentIDsByGroup(ctx, ids, params.groupID)
+			if err != nil {
+				return nil, 0, err
+			}
+			if len(filtered) == 0 {
+				return []*users.Student{}, 0, nil
+			}
+			ids = filtered
+		}
+		params.studentIDs = ids
+		// fall through to the standard ListWithOptions path. params.groupID is
+		// intentionally NOT cleared here even though buildBaseFilter ignores it
+		// because the room and group intersection was already computed via FindByIDs
+		// above, so re-applying group_id downstream would be redundant.
+	} else if params.groupID > 0 && params.locationState == "" && params.canUseGroupOnlyShortcut() {
+		// Fast path for true group-only requests keeps existing behavior.
+		students, err := rs.PersonService.GetStudentsByGroupIDs(ctx, []int64{params.groupID})
+		if err != nil {
+			return nil, 0, err
+		}
+		return students, len(students), nil
+	} else if params.groupID > 0 && params.locationState == "" {
+		students, err := rs.PersonService.GetStudentsByGroupIDs(ctx, []int64{params.groupID})
+		if err != nil {
+			return nil, 0, err
+		}
+		if len(students) == 0 {
+			return []*users.Student{}, 0, nil
+		}
+		ids := make([]int64, 0, len(students))
+		for _, student := range students {
+			if student != nil {
+				ids = append(ids, student.ID)
+			}
+		}
+		if len(ids) == 0 {
+			return []*users.Student{}, 0, nil
+		}
+		params.studentIDs = ids
+	}
+
+	// Standard path. buildBaseFilter picks up params.studentIDs (if set by
+	// the room_id branch above) and combines it with school_class /
+	// guardian_name and pagination.
+	queryOptions := params.buildQueryOptions()
+	countOptions := params.buildCountOptions()
+
+	totalCount, err := rs.StudentService.CountWithOptions(ctx, countOptions)
+	if err != nil {
+		return nil, 0, err
+	}
+
+	students, err := rs.StudentService.ListWithOptions(ctx, queryOptions)
+	if err != nil {
+		return nil, 0, err
+	}
+
+	return students, totalCount, nil
+}
+
+func (rs *Resource) listSchoolClasses(w http.ResponseWriter, r *http.Request) {
+	classes, err := rs.StudentService.ListSchoolClasses(r.Context())
+	if err != nil {
+		renderError(w, r, common.ErrorInternalServer(err))
+		return
+	}
+	common.Respond(w, r, http.StatusOK, classes, "School classes retrieved successfully")
+}
+
+// getStudent handles getting a student by ID
+func (rs *Resource) getStudent(w http.ResponseWriter, r *http.Request) {
+	student, ok := rs.parseAndGetStudent(w, r)
+	if !ok {
+		return
+	}
+
+	person, ok := rs.getPersonForStudent(w, r, student)
+	if !ok {
+		return
+	}
+
+	group := rs.getStudentGroup(r.Context(), student)
+	hasFullAccess := rs.checkStudentReadAccess(r, student)
+	hasWriteAccess := rs.checkStudentFullAccess(r, student)
+
+	attendanceLogEnabled := configService.ResolveBoolOrDefault(r.Context(), rs.SettingsService, configModel.KeyAttendanceLogEnabled, false, rs.Logger)
+	feedbackEnabled := configService.ResolveBoolOrDefault(r.Context(), rs.SettingsService, configModel.KeyFeedbackEnabled, false, rs.Logger)
+	photosEnabled := configService.ResolveBoolOrDefault(r.Context(), rs.SettingsService, configModel.KeyStudentPhotosEnabled, false, rs.Logger)
+
+	response := StudentDetailResponse{
+		StudentResponse: newStudentResponseWithOpts(r.Context(), StudentResponseOpts{
+			Student:       student,
+			Person:        person,
+			Group:         group,
+			HasFullAccess: hasFullAccess,
+			PhotosEnabled: photosEnabled,
+		}, StudentResponseServices{
+			ActiveService: rs.ActiveService,
+			PersonService: rs.PersonService,
+		}),
+		HasFullAccess:        hasFullAccess,
+		HasWriteAccess:       hasWriteAccess,
+		AttendanceLogEnabled: attendanceLogEnabled,
+		FeedbackEnabled:      feedbackEnabled,
+	}
+	now := rs.Now()
+	rs.applyStatusDaysForDateToResponse(r.Context(), &response.StudentResponse, now)
+
+	if hasFullAccess {
+		attendanceStatus, err := rs.ActiveService.GetStudentAttendanceStatus(r.Context(), student.ID)
+		if err != nil {
+			rs.Logger.Warn("failed to resolve actual student arrival/pickup times",
+				"student_id", student.ID,
+				"error", err.Error(),
+			)
+		} else {
+			applyActualTimesFromAttendance(&response.StudentResponse, attendanceStatus)
+		}
+
+		single := []StudentResponse{response.StudentResponse}
+		if err := rs.enrichWithDayPlanning(r.Context(), single, now, map[int64]*activeService.AttendanceStatus{
+			student.ID: attendanceStatus,
+		}); err != nil {
+			renderError(w, r, common.ErrorInternalServer(err))
+			return
+		}
+		response.StudentResponse = single[0]
+	}
+
+	// Add supervisor contacts for users without full access
+	if !hasFullAccess && group != nil {
+		response.GroupSupervisors = rs.buildSupervisorContacts(r.Context(), group.ID)
+	}
+
+	common.Respond(w, r, http.StatusOK, response, "Student retrieved successfully")
+}
+
+// createPersonFromStudentRequest creates a Person object from a StudentRequest
+func createPersonFromStudentRequest(req *StudentRequest) (*users.Person, error) {
+	person := &users.Person{
+		FirstName: req.FirstName,
+		LastName:  req.LastName,
+	}
+
+	// Set optional TagID if provided
+	if req.TagID != "" {
+		tagID := req.TagID
+		person.TagID = &tagID
+	}
+
+	// Set optional Birthday if provided
+	if req.Birthday != "" {
+		parsedBirthday, err := timezone.ParseDate(req.Birthday)
+		if err != nil {
+			return nil, fmt.Errorf("invalid birthday format, expected YYYY-MM-DD: %w", err)
+		}
+		person.Birthday = &parsedBirthday
+	}
+
+	return person, nil
+}
+
+// createStudentFromRequest creates a Student object from a StudentRequest and personID
+func createStudentFromRequest(req *StudentRequest, personID int64) *users.Student {
+	student := &users.Student{
+		PersonID:    personID,
+		SchoolClass: req.SchoolClass,
+	}
+
+	// Set optional legacy guardian fields if provided
+	if req.GuardianName != "" {
+		name := req.GuardianName
+		student.GuardianName = &name
+	}
+	if req.GuardianContact != "" {
+		contact := req.GuardianContact
+		student.GuardianContact = &contact
+	}
+	if req.GuardianEmail != "" {
+		email := req.GuardianEmail
+		student.GuardianEmail = &email
+	}
+	if req.GuardianPhone != "" {
+		phone := req.GuardianPhone
+		student.GuardianPhone = &phone
+	}
+
+	if req.GroupID != nil {
+		student.GroupID = req.GroupID
+	}
+	if req.AddressStreet != "" {
+		student.AddressStreet = strutil.TrimToNil(req.AddressStreet)
+	}
+	if req.AddressCity != "" {
+		student.AddressCity = strutil.TrimToNil(req.AddressCity)
+	}
+	if req.AddressPostalCode != "" {
+		student.AddressPostalCode = strutil.TrimToNil(req.AddressPostalCode)
+	}
+	if req.ExtraInfo != nil {
+		student.ExtraInfo = req.ExtraInfo
+	}
+	if req.HealthInfo != nil {
+		student.HealthInfo = req.HealthInfo
+	}
+	if req.SupervisorNotes != nil {
+		student.SupervisorNotes = req.SupervisorNotes
+	}
+	if req.DepartureCompanionNote != nil {
+		student.DepartureCompanionNote = req.DepartureCompanionNote
+	}
+	applyDeparturePlan(req.AllowedDepartureModes, req.DepartureDays, req.PickupStatus, req.PickupDays, req.Bus, req.BusDays, student)
+	normalizeDepartureCompanionNote(student)
+
+	return student
+}
+
+// normalizeDepartureCompanionNote drops the free-text "mit wem" note once the
+// child's allowed departure modes no longer include the accompanied mode, so a
+// note never outlives the "Mit anderem Kind" plan that justifies it (#1694).
+// The UI hides the note input when no day is accompanied, so a stale value can
+// otherwise sit in form state and be submitted unchanged.
+func normalizeDepartureCompanionNote(student *users.Student) {
+	if student.DepartureCompanionNote == nil {
+		return
+	}
+	if !student.AllowedDepartureModes.HasMode(users.DepartureAccompanied) &&
+		!student.DepartureDays.HasMode(users.DepartureAccompanied) {
+		student.DepartureCompanionNote = nil
+	}
+}
+
+// applyDeparturePlan sets how a child leaves each weekday from a create/update
+// request. allowed_departure_modes is the rich source of truth when present.
+// Legacy fields are passed through without rebuilding allowed_departure_modes,
+// so stale older clients cannot collapse a stored multi-mode plan before the
+// repository compares against current state.
+func applyDeparturePlan(allowed *users.AllowedDepartureModes, departure *users.DepartureDays, status *string, pickupDays *users.PickupDays, legacyBus *bool, busDays *users.BusDays, student *users.Student) {
+	if allowed == nil && departure == nil && status == nil && pickupDays == nil && legacyBus == nil && busDays == nil {
+		return
+	}
+	if allowed != nil {
+		modes := allowed.Normalize()
+		student.AllowedDepartureModes = modes
+		student.DepartureDays = modes.DepartureDays()
+		student.BusDays = modes.BusDays()
+		student.PickupDays = modes.PickupDays()
+		// Full set, not the exclusive DepartureDays() projection: bus outranks
+		// accompanied there, so a bus+accompanied day would bucket the child as a
+		// self-goer. The repository re-derives this on persist; keep the handler
+		// consistent so the in-memory student is never momentarily wrong (#1694).
+		s := modes.LegacyPickupStatus()
+		student.PickupStatus = &s
+		return
+	}
+	if departure != nil {
+		dd := departure.Normalize()
+		student.DepartureDays = dd
+		return
+	}
+	reconcilePickupFields(student, status, pickupDays)
+	applyBusDays(legacyBus, busDays, student)
+}
+
+// applyBusDays sets the student's bus_days from a create/update request.
+// bus_days is the single source of truth (#1582); the legacy bus boolean is
+// accepted only as an alias (true => Mon–Fri, false => no days) and is ignored
+// when bus_days is also supplied. The derived bus flag is no longer stored.
+func applyBusDays(legacyBus *bool, days *users.BusDays, student *users.Student) {
+	if days != nil {
+		student.BusDays = *days
+		return
+	}
+	if legacyBus == nil {
+		return
+	}
+	switch {
+	case !*legacyBus:
+		// Explicitly off: clear all bus days.
+		student.BusDays = users.BusDays{}
+	case !student.BusDays.HasAny():
+		// On with no existing per-day selection: default to all weekdays.
+		student.BusDays = users.BusDaysFromLegacyFlag(true)
+		// On with an existing per-day selection: preserve it (a legacy bus=true
+		// must not flatten Mo/Fr into all weekdays).
+	}
+}
+
+// reconcilePickupFields keeps student.PickupDays (the authoritative per-weekday
+// map) and the legacy student.PickupStatus string in sync from a request that
+// may carry either, both, or neither. When both are present the weekday map
+// wins, since it is the granular source of truth.
+//
+// Contract for legacy (status-only) callers: pickup_status is treated as
+// authoritative. A non-"Wird abgeholt" status therefore CLEARS any existing
+// weekday map — a partial update that sends pickup_status without pickup_days
+// will wipe previously stored pickup days. This is intentional (the legacy
+// string has no per-day information to preserve), but it means new clients must
+// always send pickup_days, never pickup_status alone, to mutate the map.
+func reconcilePickupFields(student *users.Student, status *string, days *users.PickupDays) {
+	if status != nil {
+		student.PickupStatus = status
+		if *status != users.PickupStatusPickedUp {
+			student.PickupDays = users.PickupDays{}
+		} else if !student.PickupDays.HasAny() {
+			student.PickupDays = users.PickupDaysFromLegacyStatus(*status)
+		}
+	}
+	if days != nil {
+		student.PickupDays = *days
+		s := student.PickupDays.LegacyPickupStatus()
+		student.PickupStatus = &s
+	}
+}
+
+// createStudent handles creating a new student with their person record
+func (rs *Resource) createStudent(w http.ResponseWriter, r *http.Request) {
+	// Parse request
+	req := &StudentRequest{}
+	if err := render.Bind(r, req); err != nil {
+		renderError(w, r, common.ErrorInvalidRequest(err))
+		return
+	}
+
+	// Create person from request
+	person, err := createPersonFromStudentRequest(req)
+	if err != nil {
+		renderError(w, r, common.ErrorInvalidRequest(err))
+		return
+	}
+
+	// Create person and student in tenant transaction
+	student := createStudentFromRequest(req, 0) // personID set after create
+
+	guardians := guardiansAPI.ToNewStudentGuardians(req.Guardians)
+
+	// Resolve the acting staff once — weekly schedules are stamped with
+	// CreatedBy. Only required when schedules are supplied so plain student
+	// creation (no schedules) is unaffected.
+	//
+	// Creating a student is governed by users:create, but attached weekly
+	// schedules are writes to the same Betreuungszeiten records edited by the
+	// standalone PUT endpoints. Keep that schedule write contract aligned:
+	// callers need users:update and must resolve to a staff record so schedule
+	// rows always carry a valid author.
+	var staffID int64
+	if len(req.ArrivalSchedules) > 0 || len(req.PickupSchedules) > 0 {
+		if !authorize.HasPermission(permissions.UsersUpdate, jwt.PermissionsFromCtx(r.Context())) {
+			renderError(w, r, common.ErrorForbidden(errors.New("users:update permission required to create student schedules")))
+			return
+		}
+		staffID, err = rs.getStaffIDFromJWT(r)
+		if err != nil {
+			renderError(w, r, common.ErrorForbidden(err))
+			return
+		}
+	}
+
+	tenantID := tenant.FromContext(r.Context())
+	if err := tenant.WithTenantTx(r.Context(), rs.DB, tenantID, func(ctx context.Context, _ bun.Tx) error {
+		// Validate guardians BEFORE writing the student. This route runs inside
+		// TenantTxMiddleware, which only rolls back on 5xx; a guardian
+		// ValidationError renders 400, so the middleware would otherwise commit
+		// an already-created student. Validating first means a 400 commits an
+		// empty transaction — no orphaned student/person rows.
+		if len(guardians) > 0 {
+			if err := rs.GuardianService.ValidateNewGuardians(ctx, guardians); err != nil {
+				return err
+			}
+		}
+
+		// Create person - validation occurs at the model layer
+		if err := rs.PersonService.Create(ctx, person); err != nil {
+			return err
+		}
+
+		// Create student with the person ID
+		student.PersonID = person.ID
+		if err := rs.StudentService.Create(ctx, student); err != nil {
+			rs.cleanupPersonAfterStudentFailure(ctx, person.ID)
+			return err
+		}
+
+		// Create any guardians supplied with the request inside the same
+		// transaction so the student and its guardians are persisted
+		// atomically — a guardian failure rolls back the whole student.
+		if len(guardians) > 0 {
+			if err := rs.GuardianService.AddGuardiansToStudent(ctx, student.ID, guardians); err != nil {
+				return err
+			}
+		}
+
+		// Persist weekly arrival/pickup schedules in the same transaction so the
+		// student and its recurring care times are created atomically (mirrors
+		// the guardian handling above). The schedule tables FK to the student,
+		// which now exists within this transaction.
+		if len(req.ArrivalSchedules) > 0 {
+			arrivals := toArrivalScheduleModels(req.ArrivalSchedules, student.ID, staffID)
+			if err := rs.ArrivalScheduleService.UpsertBulkStudentArrivalSchedules(ctx, student.ID, arrivals); err != nil {
+				return err
+			}
+		}
+		if len(req.PickupSchedules) > 0 {
+			pickups := toPickupScheduleModels(req.PickupSchedules, student.ID, staffID)
+			if err := rs.PickupScheduleService.UpsertBulkStudentPickupSchedules(ctx, student.ID, pickups); err != nil {
+				return err
+			}
+		}
+		return nil
+	}); err != nil {
+		// Bad guardian input (e.g. invalid email) is a client error: the
+		// transaction has already rolled back, so no partial data survives.
+		var validationErr *userService.ValidationError
+		if errors.As(err, &validationErr) {
+			renderError(w, r, common.ErrorInvalidRequest(validationErr))
+			return
+		}
+		renderError(w, r, common.ErrorInternalServer(err))
+		return
+	}
+
+	// Get group data if student has a group
+	group := rs.fetchStudentGroup(r.Context(), student.GroupID)
+
+	// Admin users creating students can see full data including detailed location
+	userPermissions := jwt.PermissionsFromCtx(r.Context())
+	hasFullAccess := authorize.HasAdminWildcard(userPermissions)
+
+	// Return the created student with person data
+	photosEnabled := configService.ResolveBoolOrDefault(r.Context(), rs.SettingsService, configModel.KeyStudentPhotosEnabled, false, rs.Logger)
+	common.Respond(w, r, http.StatusCreated, newStudentResponseWithOpts(r.Context(), StudentResponseOpts{
+		Student:       student,
+		Person:        person,
+		Group:         group,
+		HasFullAccess: hasFullAccess,
+		PhotosEnabled: photosEnabled,
+	}, StudentResponseServices{
+		ActiveService: rs.ActiveService,
+		PersonService: rs.PersonService,
+	}), "Student created successfully")
+}
+
+// cleanupPersonAfterStudentFailure removes the person record if student creation fails
+func (rs *Resource) cleanupPersonAfterStudentFailure(ctx context.Context, personID int64) {
+	if err := rs.PersonService.Delete(ctx, personID); err != nil {
+		slog.Default().Error("failed to cleanup person after failed student creation",
+			slog.Int64("person_id", personID),
+			slog.String("error", err.Error()))
+	}
+}
+
+// personUpdateResult contains the result of updating person fields
+type personUpdateResult struct {
+	updated bool
+	err     error
+}
+
+// applyPersonUpdates applies person field changes from the request
+// Returns whether any fields were updated and any error encountered
+func applyPersonUpdates(req *UpdateStudentRequest, person *users.Person) personUpdateResult {
+	result := personUpdateResult{}
+
+	if req.FirstName != nil {
+		person.FirstName = *req.FirstName
+		result.updated = true
+	}
+	if req.LastName != nil {
+		person.LastName = *req.LastName
+		result.updated = true
+	}
+	if req.Birthday != nil {
+		if *req.Birthday != "" {
+			parsedBirthday, err := timezone.ParseDate(*req.Birthday)
+			if err != nil {
+				result.err = fmt.Errorf("invalid birthday format, expected YYYY-MM-DD: %w", err)
+				return result
+			}
+			person.Birthday = &parsedBirthday
+		} else {
+			person.Birthday = nil
+		}
+		result.updated = true
+	}
+	if req.TagID != nil {
+		if *req.TagID != "" {
+			person.TagID = req.TagID
+		} else {
+			person.TagID = nil
+		}
+		result.updated = true
+	}
+
+	return result
+}
+
+// applyStudentFieldUpdates applies student field changes from the request
+func applyStudentFieldUpdates(req *UpdateStudentRequest, student *users.Student) {
+	if req.SchoolClass != nil {
+		student.SchoolClass = *req.SchoolClass
+	}
+	applyGuardianUpdates(req, student)
+	applyOptionalStudentFields(req, student)
+	applySickStatus(req, student)
+	applyExcusedStatus(req, student)
+}
+
+func reconcilePhotoConsentRequest(requested *bool, snapshot, fresh *users.Student) *bool {
+	if requested == nil {
+		return nil
+	}
+	snapshotHadConsent := snapshot != nil && snapshot.PhotoConsentGivenAt != nil
+	freshHasConsent := fresh != nil && fresh.PhotoConsentGivenAt != nil
+
+	// Treat values that merely echo the pre-transaction snapshot as no-ops.
+	// Old clients used to serialize photo_consent_given on every PUT; if another
+	// tab changed consent between the snapshot read and this row lock, replaying
+	// that stale unchanged boolean would re-grant withdrawn consent or withdraw a
+	// newly granted consent/photo.
+	if *requested == snapshotHadConsent {
+		return nil
+	}
+
+	// If a concurrent request already completed the intended transition, do not
+	// re-stamp audit metadata or schedule duplicate photo cleanup.
+	if *requested == freshHasConsent {
+		return nil
+	}
+
+	return requested
+}
+
+// applyGuardianUpdates handles legacy guardian field updates
+func applyGuardianUpdates(req *UpdateStudentRequest, student *users.Student) {
+	if req.GuardianName != nil {
+		trimmed := strings.TrimSpace(*req.GuardianName)
+		if trimmed == "" {
+			student.GuardianName = nil
+		} else {
+			student.GuardianName = &trimmed
+		}
+	}
+	if req.GuardianContact != nil {
+		trimmed := strings.TrimSpace(*req.GuardianContact)
+		if trimmed == "" {
+			student.GuardianContact = nil
+		} else {
+			student.GuardianContact = &trimmed
+		}
+	}
+	if req.GuardianEmail != nil {
+		student.GuardianEmail = req.GuardianEmail
+	}
+	if req.GuardianPhone != nil {
+		student.GuardianPhone = req.GuardianPhone
+	}
+}
+
+// applyOptionalStudentFields applies optional fields like GroupID, ExtraInfo, etc.
+func applyOptionalStudentFields(req *UpdateStudentRequest, student *users.Student) {
+	if req.GroupID != nil {
+		student.GroupID = req.GroupID
+	}
+	if req.AddressStreet != nil {
+		student.AddressStreet = strutil.TrimToNil(*req.AddressStreet)
+	}
+	if req.AddressCity != nil {
+		student.AddressCity = strutil.TrimToNil(*req.AddressCity)
+	}
+	if req.AddressPostalCode != nil {
+		student.AddressPostalCode = strutil.TrimToNil(*req.AddressPostalCode)
+	}
+	if req.ExtraInfo != nil {
+		student.ExtraInfo = req.ExtraInfo
+	}
+	if req.HealthInfo != nil {
+		student.HealthInfo = req.HealthInfo
+	}
+	if req.SupervisorNotes != nil {
+		student.SupervisorNotes = req.SupervisorNotes
+	}
+	if req.DepartureCompanionNote != nil {
+		student.DepartureCompanionNote = req.DepartureCompanionNote
+	}
+	applyDeparturePlan(req.AllowedDepartureModes, req.DepartureDays, req.PickupStatus, req.PickupDays, req.Bus, req.BusDays, student)
+	// Only normalize when this request actually set the departure plan, so the
+	// freshly applied modes are authoritative; an update that omits modes must
+	// not clear a note against a possibly-unpopulated scanonly field.
+	if req.AllowedDepartureModes != nil || req.DepartureDays != nil {
+		normalizeDepartureCompanionNote(student)
+	}
+}
+
+// applySickStatus handles sick status updates with SickSince timestamp logic
+func applySickStatus(req *UpdateStudentRequest, student *users.Student) {
+	if req.Sick == nil {
+		return
+	}
+	student.Sick = req.Sick
+	if *req.Sick {
+		if student.SickSince == nil {
+			now := time.Now()
+			student.SickSince = &now
+		}
+	} else {
+		student.SickSince = nil
+	}
+}
+
+// applyExcusedStatus handles excused status updates with ExcusedSince timestamp logic
+func applyExcusedStatus(req *UpdateStudentRequest, student *users.Student) {
+	if req.Excused == nil {
+		return
+	}
+	student.Excused = req.Excused
+	if *req.Excused {
+		if student.ExcusedSince == nil {
+			now := time.Now()
+			student.ExcusedSince = &now
+		}
+	} else {
+		student.ExcusedSince = nil
+	}
+}
+
+// checkSickExcusedConflict returns an error if the incoming update would
+// result in both sick and excused being true simultaneously. Callers with a
+// conflict should prompt the user to switch states rather than hold both.
+func checkSickExcusedConflict(req *UpdateStudentRequest, student *users.Student) error {
+	sickFinal := student.Sick != nil && *student.Sick
+	if req.Sick != nil {
+		sickFinal = *req.Sick
+	}
+	excusedFinal := student.Excused != nil && *student.Excused
+	if req.Excused != nil {
+		excusedFinal = *req.Excused
+	}
+	if sickFinal && excusedFinal {
+		return errors.New("a student cannot be both sick and excused at the same time")
+	}
+	return nil
+}
+
+// broadcastStudentUpdated emits an SSE student_updated event to the
+// AFFECTED tenant's connected clients only. use-global-sse.ts treats
+// that event as "invalidate every student / room / dashboard cache in
+// this tab", so a global fan-out (BroadcastToAll) would force tabs in
+// schools B to N to refetch unrelated data whenever school A edits one
+// student or one avatar. Routing via BroadcastToTenant, the helper
+// already added in this branch for tenant_settings_changed, keeps the
+// invalidation scoped to the school that actually changed.
+//
+// Callers MUST pass tenantID from request context (tenant.FromContext)
+// or from a captured value inside an after-commit hook. Passing zero
+// no-ops the broadcast rather than fanning out, since
+// BroadcastToTenant rejects zero tenant IDs by definition (no Client
+// has TenantID == 0).
+func (rs *Resource) broadcastStudentUpdated(tenantID, studentID int64) {
+	if rs.Broadcaster == nil {
+		return
+	}
+	if tenantID <= 0 {
+		// Defensive: a missing tenant context means we don't know which
+		// school's clients should invalidate. Logging the case so a
+		// future caller that forgets to thread tenantID through gets a
+		// breadcrumb instead of silent loss.
+		if rs.Logger != nil {
+			rs.Logger.Warn(
+				"skipping student_updated broadcast, no tenant context",
+				"student_id", studentID,
+			)
+		}
+		return
+	}
+
+	source := "manual"
+	event := realtime.NewEvent(realtime.EventStudentUpdated, "", realtime.EventData{
+		Source: &source,
+	})
+
+	if err := rs.Broadcaster.BroadcastToTenant(tenantID, event); err != nil && rs.Logger != nil {
+		rs.Logger.Warn(
+			"failed to broadcast student update",
+			"tenant_id", tenantID,
+			"student_id", studentID,
+			"error", err.Error(),
+		)
+	}
+}
+
+// updateStudent handles updating an existing student
+func (rs *Resource) updateStudent(w http.ResponseWriter, r *http.Request) {
+	// Parse ID and get student
+	student, ok := rs.parseAndGetStudent(w, r)
+	if !ok {
+		return
+	}
+
+	// Parse request
+	req := &UpdateStudentRequest{}
+	if err := render.Bind(r, req); err != nil {
+		renderError(w, r, common.ErrorInvalidRequest(err))
+		return
+	}
+
+	// Get existing person
+	person, ok := rs.getPersonForStudent(w, r, student)
+	if !ok {
+		return
+	}
+
+	// Centralized permission check for updating student data
+	userPermissions := jwt.PermissionsFromCtx(r.Context())
+	authorized, authErr := canUpdateStudent(r.Context(), userPermissions, student, rs.UserContextService)
+	if !authorized {
+		renderError(w, r, common.ErrorForbidden(authErr))
+		return
+	}
+
+	// Track whether the user is admin or group supervisor
+	isAdmin := authorize.HasAdminWildcard(userPermissions)
+	isGroupSupervisor := !isAdmin // If not admin but authorized, must be group supervisor
+
+	// Update person fields using helper function
+	personResult := applyPersonUpdates(req, person)
+	if personResult.err != nil {
+		renderError(w, r, common.ErrorInvalidRequest(personResult.err))
+		return
+	}
+
+	// Reject updates that would leave the student in both sick and excused
+	// states simultaneously. The frontend uses the SICK_EXCUSED_CONFLICT code
+	// to prompt the user to switch states rather than hold both.
+	if err := checkSickExcusedConflict(req, student); err != nil {
+		renderError(w, r, common.ErrorConflictWithCode(err, ErrCodeSickExcusedConflict))
+		return
+	}
+
+	statusHistoryNow := time.Now()
+
+	// In-tx sentinel: a concurrent partial update committed between our
+	// pre-tx conflict check and the locked-row re-check below produced
+	// the forbidden sick && excused state. Mapped to 409 +
+	// SICK_EXCUSED_CONFLICT in the outer error switch so the frontend
+	// can prompt the user the same way it does for a synchronous
+	// conflict. errors.Is keeps the dispatch resilient to wrapping.
+	errSickExcusedConflict := errors.New("sick and excused conflict on locked row")
+
+	// In-tx sentinel: a concurrent delete removed the student row
+	// between parseAndGetStudent above and the locked re-read below.
+	// The non-racy path returns 404; bare sql.ErrNoRows would otherwise
+	// fall through to a 500 for a legitimate concurrent delete.
+	errStudentNotFoundUnderLock := errors.New("student deleted between snapshot and lock")
+
+	// In-tx sentinel: the pre-tx canUpdateStudent check ran on
+	// student.GroupID from the snapshot. canUpdateStudent decides off
+	// group membership, so a concurrent admin moving the student into
+	// a different group between snapshot and lock can leave the caller
+	// without write authority on the locked row. Re-checking against
+	// fresh closes that window. Mapped to 403 in the outer switch so
+	// the response status matches what the pre-tx gate would emit.
+	errStudentReassigned := errors.New("student reassigned out of caller's scope mid-update")
+
+	// Patch is applied to a freshly FOR-UPDATE-locked row inside the tx so a
+	// concurrent photo upload can't have its photo_path clobbered by a stale
+	// snapshot write.
+	tenantID := tenant.FromContext(r.Context())
+	if err := tenant.WithTenantTx(r.Context(), rs.DB, tenantID, func(ctx context.Context, _ bun.Tx) error {
+		// Acquire the photo-feature advisory lock only when consent is
+		// actually toggling. Name/notes edits must not queue behind
+		// feature disable/purge.
+		consentChanging := req.PhotoConsentGiven != nil &&
+			(*req.PhotoConsentGiven) != (student.PhotoConsentGivenAt != nil)
+		if consentChanging {
+			if err := rs.StudentService.LockPhotoFeature(ctx); err != nil {
+				return err
+			}
+		}
+
+		fresh, err := rs.StudentService.GetByIDForUpdate(ctx, student.ID)
+		if err != nil {
+			if errors.Is(err, sql.ErrNoRows) {
+				return errStudentNotFoundUnderLock
+			}
+			return err
+		}
+
+		// Re-check authorisation against the LOCKED row. A concurrent admin
+		// reassignment could otherwise let a non-supervisor mutate the row.
+		if ok, _ := canUpdateStudent(ctx, userPermissions, fresh, rs.UserContextService); !ok {
+			return errStudentReassigned
+		}
+
+		// Re-validate sick/excused on fresh: two concurrent partial updates
+		// against a stale snapshot can otherwise merge into the forbidden
+		// sick && excused state. TenantTxMiddleware only rolls back on 5xx,
+		// so this 409-mapped sentinel must fire before any write below.
+		if err := checkSickExcusedConflict(req, fresh); err != nil {
+			return errSickExcusedConflict
+		}
+
+		if personResult.updated {
+			if err := rs.PersonService.Update(ctx, person); err != nil {
+				return err
+			}
+		}
+
+		// Read pre-update flags off fresh, not the snapshot, so status
+		// history reflects the actual transition the commit will perform.
+		wasSick := boolPtrValue(fresh.Sick)
+		wasExcused := boolPtrValue(fresh.Excused)
+
+		applyStudentFieldUpdates(req, fresh)
+		effectiveConsent := reconcilePhotoConsentRequest(req.PhotoConsentGiven, student, fresh)
+		rs.StudentPhotos.ApplyConsentTransition(ctx, effectiveConsent, fresh)
+
+		if err := rs.persistStudentStatusHistory(ctx, fresh, wasSick, wasExcused, statusHistoryNow, strutil.TrimPtrToNil(req.SickReason)); err != nil {
+			rs.logStatusHistoryError(student.ID, err)
+			return err
+		}
+		if err := rs.StudentService.Update(ctx, fresh); err != nil {
+			return err
+		}
+
+		// Broadcast after the OUTER tx commits. Broadcasting now would race
+		// subscribers into refetching the still-pre-commit row.
+		studentID := student.ID
+		capturedTenantID := tenantID
+		tenant.RegisterAfterCommit(ctx, func() {
+			rs.broadcastStudentUpdated(capturedTenantID, studentID)
+		})
+		return nil
+	}); err != nil {
+		if errors.Is(err, errSickExcusedConflict) {
+			renderError(w, r, common.ErrorConflictWithCode(
+				errors.New("a student cannot be both sick and excused at the same time"),
+				ErrCodeSickExcusedConflict,
+			))
+			return
+		}
+		if errors.Is(err, errStudentReassigned) {
+			renderError(w, r, common.ErrorForbidden(errors.New("you can only update students in groups you supervise")))
+			return
+		}
+		if errors.Is(err, errStudentNotFoundUnderLock) {
+			renderError(w, r, common.ErrorNotFound(errors.New("student not found")))
+			return
+		}
+		// The merged plan (request modes applied onto the stored row) can violate
+		// the accompanied-requires-note invariant — e.g. a caller sets a "Mit
+		// anderem Kind" day on a child with no stored note. That is client input,
+		// so surface it as a 400 rather than the model error leaking as a 500
+		// (#1694). The binder cannot catch this on update: only here is the
+		// stored note visible to fall back on.
+		if errors.Is(err, users.ErrDepartureCompanionNoteRequired) {
+			renderError(w, r, common.ErrorInvalidRequest(err))
+			return
+		}
+		renderError(w, r, common.ErrorInternalServer(err))
+		return
+	}
+
+	// Get updated student with person data
+	updatedStudent, err := rs.PersonService.GetStudentByID(r.Context(), student.ID)
+	if err != nil {
+		renderError(w, r, common.ErrorInternalServer(err))
+		return
+	}
+
+	// Get group data if student has a group
+	group := rs.getStudentGroup(r.Context(), updatedStudent)
+
+	// Admin users and group supervisors can see full data including detailed location
+	// Explicitly verify access level based on the checks performed above
+	hasFullAccess := isAdmin || isGroupSupervisor // Explicitly check for admin or group supervisor
+
+	// Return the updated student with person data
+	photosEnabled := configService.ResolveBoolOrDefault(r.Context(), rs.SettingsService, configModel.KeyStudentPhotosEnabled, false, rs.Logger)
+	common.Respond(w, r, http.StatusOK, newStudentResponseWithOpts(r.Context(), StudentResponseOpts{
+		Student:       updatedStudent,
+		Person:        person,
+		Group:         group,
+		HasFullAccess: hasFullAccess,
+		PhotosEnabled: photosEnabled,
+	}, StudentResponseServices{
+		ActiveService: rs.ActiveService,
+		PersonService: rs.PersonService,
+	}), "Student updated successfully")
+}
+
+// deleteStudent handles deleting a student and their associated person record
+func (rs *Resource) deleteStudent(w http.ResponseWriter, r *http.Request) {
+	// Parse ID and get student
+	student, ok := rs.parseAndGetStudent(w, r)
+	if !ok {
+		return
+	}
+
+	// Check if user has permission to delete this student
+	userPermissions := jwt.PermissionsFromCtx(r.Context())
+	authorized, authErr := canDeleteStudent(r.Context(), userPermissions, student, rs.UserContextService)
+	if !authorized {
+		renderError(w, r, common.ErrorForbidden(authErr))
+		return
+	}
+
+	// Capture the photo path from the locked row (not the pre-tx snapshot)
+	// so a concurrent upload can't orphan a new file on disk. The unlink
+	// itself runs after the OUTER tenant tx commits.
+	tenantID := tenant.FromContext(r.Context())
+	if err := tenant.WithTenantTx(r.Context(), rs.DB, tenantID, func(ctx context.Context, _ bun.Tx) error {
+		// FOR UPDATE row-locks against any in-flight upload tx. We either
+		// observe its committed photo_path or it sees our deleted row and
+		// aborts.
+		fresh, err := rs.StudentService.GetByIDForUpdate(ctx, student.ID)
+		if err != nil {
+			return err
+		}
+		var photoToRemove string
+		if fresh.PhotoPath != nil {
+			photoToRemove = *fresh.PhotoPath
+		}
+
+		if err := rs.StudentService.Delete(ctx, student.ID); err != nil {
+			return err
+		}
+
+		// Person delete failure must not fail the request. The student row
+		// is already gone, leaving the person orphaned is recoverable.
+		if err := rs.PersonService.Delete(ctx, student.PersonID); err != nil {
+			slog.Default().Error("failed to delete associated person record",
+				slog.Int64("person_id", student.PersonID),
+				slog.String("error", err.Error()))
+		}
+
+		rs.StudentPhotos.ScheduleUnlinkAfterCommit(ctx, photoToRemove)
+		return nil
+	}); err != nil {
+		if common.IsConstraintViolation(err) {
+			renderError(w, r, common.ErrorConflictMessage("Kind kann nicht gelöscht werden: Kind hat aktive Besuche, Einschreibungen oder andere verknüpfte Daten"))
+			return
+		}
+		renderError(w, r, common.ErrorInternalServer(err))
+		return
+	}
+
+	common.Respond(w, r, http.StatusOK, nil, "Student deleted successfully")
+}
