+++ backend/test/handler_complexity_ratchet_test.go
@@ -36,10 +36,10 @@ const handlerComplexityThreshold = 15
 // Seeded 2026-07-12 from `gocognit -over 15 api/` (non-test files only), the
 // issue #575 B0 baseline. 75 functions.
 var handlerComplexityAllowlist = map[string]int{
-	"api/activities/api.go:buildReplacementSupervisorIDs":                                       22,
-	"api/auth/invitation_handlers.go:(*Resource).createInvitation":                              20,
-	"api/base.go:(*API).registerRoutesWithRateLimiting":                                         20,
-	"api/base.go:setupCORS":                                                                     18,
+	"api/activities/supervisor_handlers.go:buildReplacementSupervisorIDs": 22,
+	"api/auth/invitation_handlers.go:(*Resource).createInvitation":        20,
+	"api/base.go:(*API).registerRoutesWithRateLimiting":                   20,
+	"api/base.go:setupCORS": 18,
 	"api/common/student_locations.go:(*StudentLocationSnapshot).ResolveStudentLocationWithTime": 16,
 	"api/enrollment/admin_decision_handlers.go:(*Resource).decideAdminChild":                    20,
 	"api/enrollment/admin_decision_handlers.go:(*Resource).toAdminRequestDetailSummary":         21,
@@ -69,13 +69,13 @@ var handlerComplexityAllowlist = map[string]int{
 	"api/server.go:NewServer":                                                                  35,
 	"api/sse/api.go:(*Resource).eventsHandler":                                                 17,
 	"api/sse/sse_helpers.go:(*Resource).resolveSupervisions":                                   16,
-	"api/staff/api.go:(*Resource).buildScheduleResponse":                                       19,
-	"api/staff/api.go:(*Resource).getStaff":                                                    19,
-	"api/staff/api.go:(*Resource).getStaffByRole":                                              17,
-	"api/staff/api.go:(*Resource).updateSchedule":                                              20,
-	"api/students/api.go:(*Resource).createStudent":                                            35,
-	"api/students/api.go:(*Resource).fetchStudentsForList":                                     53,
-	"api/students/api.go:(*Resource).updateStudent":                                            41,
+	"api/staff/schedule_handlers.go:(*Resource).buildScheduleResponse":                         19,
+	"api/staff/staff_handlers.go:(*Resource).getStaff":                                         19,
+	"api/staff/substitution_handlers.go:(*Resource).getStaffByRole":                            17,
+	"api/staff/schedule_handlers.go:(*Resource).updateSchedule":                                20,
+	"api/students/student_handlers.go:(*Resource).createStudent":                               35,
+	"api/students/student_handlers.go:(*Resource).fetchStudentsForList":                        53,
+	"api/students/student_handlers.go:(*Resource).updateStudent":                               41,
 	"api/students/arrival_schedule_handlers.go:(*Resource).createStudentArrivalException":      19,
 	"api/students/arrival_schedule_handlers.go:(*Resource).updateStudentArrivalException":      24,
 	"api/students/attendance_history_handlers.go:(*Resource).getStudentAttendanceHistory":      20,
